/admin/products-controller.js - admin/order-controller.js - admin/category-controller.js - admin/brand-controller.js - shop/products-controller.js - shop/order-controller.js - buradaki mongoose kontrolcusu sonrasi hafif bu sepet kisminda yavaslama oldu urun arttirma azaltma ve eksiltem

+++ server/controllers/admin/products-controller.js
@@ -1,5 +1,6 @@
 const { imageUploadUtil } = require("../../helpers/cloudinary");
 const Product = require("../../models/Product");
+const mongoose = require("mongoose");
 
 const handleImageUpload = async (req, res) => {
   try {
@@ -86,6 +87,11 @@ const fetchAllProducts = async (req, res) => {
 const editProduct = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Geçersiz Ürün ID formatı." });
+    }
     const {
       image,
       title,
@@ -136,6 +142,11 @@ const editProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Geçersiz Ürün ID formatı." });
+    }
     const product = await Product.findByIdAndDelete(id);
 
     if (!product)
